pipelines/fast_imaging_pipeline_distributed_flink/functions/func_requirements.txt
numpy
astropy
matplotlib
pybase64
python-casacore